Signup, Login and Welcome RP
This is what I have so far for the Signup, Login, and Welcome pages. The Signup page takes the username and password inputs and puts them into the Global Variables "SignUpUsername", and "SignUpPassword". The login page will then check for these variables and if they are incorrect it will give an error message. 
If they are entered correctly, however, it will take you to the welcome screen, where it will say "Welcome," and then whatever your username is. 
The home page is literally just two buttons which lead to either signup or login so I've taken the liberty to call it "Arthur Home" in case any of you called it "Home"